requirements.txt
ultralytics>=8.3.0
torch==2.1.2
torchvision==0.16.2
opencv-python-headless==4.8.1.78
pillow
numpy==1.26.4
flask==3.0.0
